test: Reorganize Playwright settings tests to match UI structure
