+++ packages/testing/playwright/tests/ui/settings/workers/workers.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../../fixtures/base';
+import { test, expect } from '../../../../fixtures/base';
 
 test.describe
 	.serial('Worker View', () => {
